Alterations on the state struct.

+++ include/CCH/CCH_mechanics.h
@@ -112,6 +112,6 @@ int CCH_same_color(CCH_Piece pa, CCH_Piece pb);
 int CCH_move_piece(CCH_State* state, CCH_Move mv);
 
 // TODO document
-int CCH_aval_simple_pos(const CCH_State*);
+float CCH_aval_simple_pos(const CCH_State*);
 
 #endif // INCLUDE_CCH_MECHANICS_H_
